Refactored AutoComplete to separate class
