+++ src/JuliusSweetland.OptiKey/JuliusSweetland.OptiKey.csproj
@@ -255,6 +255,8 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <Compile Include="Services\AutoComplete\BasicAutoComplete.cs" />
+    <Compile Include="Services\AutoComplete\IAutoComplete.cs" />
     <Compile Include="Services\ICalibrationService.cs" />
     <Compile Include="Services\IMouseOutputService.cs" />
     <Compile Include="Services\INotifyErrors.cs" />
