+++ src/JuliusSweetland.OptiKey/Services/AutoComplete/IAutoComplete.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace JuliusSweetland.OptiKey.Services.AutoComplete
+{
+    internal interface IAutoComplete
+    {
+        IEnumerable<string> GetSuggestions(string root);
+    }
+}
